requirements.txt
requests
django-environ
mediacloud
pandas
csv